+++ src/main/java/janggi/domain/Movement.java
@@ -1,4 +1,4 @@
-package janggi;
+package janggi.domain;
 
 import java.util.List;
