+++ src/main/java/janggi/domain/Team.java
@@ -1,4 +1,4 @@
-package janggi;
+package janggi.domain;
 
 public enum Team {
     RED,
