+++ src/test/java/janggi/domain/BoardTest.java
@@ -1,13 +1,12 @@
-package janggi;
+package janggi.domain;
 
-import janggi.piece.Piece;
-import janggi.piece.Soldier;
+import janggi.domain.piece.Piece;
+import janggi.domain.piece.Soldier;
+import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
