+++ src/test/java/janggi/domain/PositionTest.java
@@ -1,4 +1,4 @@
-package janggi;
+package janggi.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
