+++ src/test/java/janggi/domain/RouteTest.java
@@ -1,10 +1,9 @@
-package janggi;
+package janggi.domain;
 
+import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
 
 public class RouteTest {
