+++ src/test/java/janggi/domain/piece/ElephantTest.java
@@ -1,8 +1,8 @@
-package janggi.piece;
+package janggi.domain.piece;
 
-import janggi.Board;
-import janggi.Position;
-import janggi.Team;
+import janggi.domain.Board;
+import janggi.domain.Position;
+import janggi.domain.Team;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,39 +12,39 @@ import org.junit.jupiter.params.provider.CsvSource;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class HorseTest {
+class ElephantTest {
 
     @Test
-    @DisplayName("마는 수직/수평으로 1칸 이동 후, 진행 방향의 대각선으로 1칸 이동할 수 있다")
+    @DisplayName("상은 수직/수평으로 1칸 이동 후, 진행 방향의 대각선으로 2칸 이동할 수 있다")
     void move() {
         // given
         Position position = Position.of(5, 5);
-        Piece horse = new Horse(position, Team.RED);
-        Board board = Board.initialize(List.of(horse));
+        Piece elephant = new Elephant(position, Team.RED);
+        Board board = Board.initialize(List.of(elephant));
 
-        Position movedPosition = position.adjust(1, 2);
+        Position movedPosition = position.adjust(2, 3);
 
         // when
-        Piece move = horse.move(board, movedPosition);
+        Piece move = elephant.move(board, movedPosition);
 
         // then
         assertThat(move.getPosition()).isEqualTo(movedPosition);
     }
 
     @ParameterizedTest
-    @CsvSource(value = {"1, 3", "3, 1", "2,2"})
-    @DisplayName("마는 규칙에 어긋나게 움직일 수 없다")
+    @CsvSource(value = {"2, 4", "4, 2", "3,3", "1,1"})
+    @DisplayName("상은 규칙에 어긋나게 움직일 수 없다")
     void move(int rowDirection, int columnDirection) {
         // given
         Position position = Position.of(5, 5);
-        Piece horse = new Horse(position, Team.RED);
-        Board board = Board.initialize(List.of(horse));
+        Piece elephant = new Elephant(position, Team.RED);
+        Board board = Board.initialize(List.of(elephant));
 
         Position movedPosition = position.adjust(rowDirection, columnDirection);
 
         // when
         // then
-        assertThatThrownBy(() -> horse.move(board, movedPosition))
+        assertThatThrownBy(() -> elephant.move(board, movedPosition))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("이동할 수 없는 지점입니다.");
     }
